// src/types.ts
export interface InfactorySDKError {
  status: number;
  message: string;
}
export interface ApiResponse<T> {
  data?: T;
  error?: InfactorySDKError;
}

// Base types
export interface BaseEntity {
  id: string;
  created_at: string;
  updated_at: string;
  deleted_at?: string | null;
}
// ...
export interface Infrastructure extends BaseEntity {
  name: string;
  organization_id: string;
  type: string;
  config: Record<string, any>;
}

export interface CreateInfrastructureParams {
  name: string;
  organization_id: string;
  type: string;
  config: Record<string, any>;
}